+++ poulpy-core/src/default/operations/glwe.rs
@@ -18,8 +18,10 @@ use crate::{
     default::keyswitching::{GGLWEProductDefault, gglwe_product_output_size},
     layouts::{
         Base2K, GGLWEInfos, GLWEInfos, GLWEToBackendMut, GLWEToBackendRef, IntPolyInfos, LWEInfos,
-        prepared::GLWETensorKeyPreparedToBackendRef,
+        prepared::{GGLWEPreparedToBackendRef, GLWETensorKeyPreparedToBackendRef},
     },
+    oep::GLWETensoringImpl,
+    scratch::ScratchArenaTakeCore,
 };
 
 #[doc(hidden)]
@@ -512,6 +514,57 @@ pub trait GLWETensoringDefault<BE: Backend> {
         R: GLWEToBackendMut<BE> + GLWEInfos,
         A: GLWEToBackendRef<BE> + GLWEInfos,
         BP: CnvPVecRToBackendRef<BE>;
+
+    /// Materializes `tensor_infos`, applies, relinearizes. Dispatches through
+    /// `BE::glwe_tensor_*` so a backend override of a suboperation is kept.
+    #[allow(clippy::too_many_arguments)]
+    fn glwe_tensor_apply_relinearize_default<R, I, A, B, T>(
+        &self,
+        cnv_offset: usize,
+        res: &mut R,
+        tensor_infos: &I,
+        a: &A,
+        b: &B,
+        tsk: &T,
+        scratch: &mut ScratchArena<'_, BE>,
+    ) where
+        BE: GLWETensoringImpl<BE>,
+        R: GLWEToBackendMut<BE> + GLWEInfos,
+        I: GLWEInfos,
+        A: GLWEToBackendRef<BE> + GLWEInfos,
+        B: GLWEToBackendRef<BE> + GLWEInfos,
+        T: GGLWEInfos + GLWETensorKeyPreparedToBackendRef<BE> + GGLWEPreparedToBackendRef<BE>;
+
+    #[allow(clippy::too_many_arguments)]
+    fn glwe_tensor_square_relinearize_assign_default<R, I, T>(
+        &self,
+        cnv_offset: usize,
+        res: &mut R,
+        tensor_infos: &I,
+        tsk: &T,
+        scratch: &mut ScratchArena<'_, BE>,
+    ) where
+        BE: GLWETensoringImpl<BE>,
+        R: GLWEToBackendMut<BE> + GLWEToBackendRef<BE> + GLWEInfos,
+        I: GLWEInfos,
+        T: GGLWEInfos + GLWETensorKeyPreparedToBackendRef<BE> + GGLWEPreparedToBackendRef<BE>;
+
+    #[allow(clippy::too_many_arguments)]
+    fn glwe_tensor_apply_prepared_right_relinearize_assign_default<R, I, BP, T>(
+        &self,
+        cnv_offset: usize,
+        res: &mut R,
+        tensor_infos: &I,
+        b_prep: &BP,
+        b_size: usize,
+        tsk: &T,
+        scratch: &mut ScratchArena<'_, BE>,
+    ) where
+        BE: GLWETensoringImpl<BE>,
+        R: GLWEToBackendMut<BE> + GLWEToBackendRef<BE> + GLWEInfos,
+        I: GLWEInfos,
+        BP: CnvPVecRToBackendRef<BE>,
+        T: GGLWEInfos + GLWETensorKeyPreparedToBackendRef<BE> + GGLWEPreparedToBackendRef<BE>;
 }
 
 impl<BE: Backend> GLWETensoringDefault<BE> for Module<BE>
@@ -627,7 +680,6 @@ where
 
         let a_base2k: usize = a.base2k().into();
         let key_base2k: usize = tsk.base2k().into();
-        let res_base2k: usize = res.base2k().into();
 
         let cols: usize = tsk.rank_out().as_usize() + 1;
         let pairs: usize = tsk.rank_in().as_usize();
@@ -637,23 +689,14 @@ where
 
         let lvl_0: usize = self.bytes_of_vec_znx_dft(pairs, a_dft_size);
 
-        let lvl_1_pre_conv: usize = if a_base2k != key_base2k {
-            BE::bytes_of_vec_znx(self.n(), 1, a_dft_size) + self.vec_znx_normalize_tmp_bytes()
-        } else {
-            0
-        };
+        // Both `a_conv` stages are taken unconditionally, whatever the radices.
+        let conv: usize = BE::bytes_of_vec_znx(self.n(), 1, a_dft_size) + self.vec_znx_normalize_tmp_bytes();
         let lvl_1_res_dft: usize = self.bytes_of_vec_znx_dft(cols, output_size);
         let lvl_1_gglwe_product: usize = self.gglwe_product_dft_tmp_bytes_default(output_size, a_dft_size, tsk);
-        let lvl_1_post_conv: usize = if res_base2k != key_base2k {
-            BE::bytes_of_vec_znx(self.n(), 1, a_dft_size) + self.vec_znx_normalize_tmp_bytes()
-        } else {
-            0
-        };
         let lvl_1_big_norm: usize = self.bytes_of_vec_znx_big(cols, output_size)
-            + BE::bytes_of_vec_znx(self.n(), 1, res.size())
-            + self.vec_znx_big_normalize_tmp_bytes();
-        let lvl_1_main: usize = lvl_1_res_dft + lvl_1_gglwe_product.max(lvl_1_post_conv).max(lvl_1_big_norm);
-        let lvl_1: usize = lvl_1_pre_conv.max(lvl_1_main);
+            + conv.max(BE::bytes_of_vec_znx(self.n(), 1, res.size()) + self.vec_znx_big_normalize_tmp_bytes());
+        let lvl_1_main: usize = lvl_1_res_dft + lvl_1_gglwe_product.max(lvl_1_big_norm);
+        let lvl_1: usize = conv.max(lvl_1_main);
 
         lvl_0 + lvl_1
     }
@@ -863,6 +906,70 @@ where
         BP: CnvPVecRToBackendRef<BE>,
     {
         glwe_tensor_apply_prepared_right(self, cnv_offset, res, a, b_prep, b_size, scratch);
+    }
+
+    fn glwe_tensor_apply_relinearize_default<R, I, A, B, T>(
+        &self,
+        cnv_offset: usize,
+        res: &mut R,
+        tensor_infos: &I,
+        a: &A,
+        b: &B,
+        tsk: &T,
+        scratch: &mut ScratchArena<'_, BE>,
+    ) where
+        BE: GLWETensoringImpl<BE>,
+        R: GLWEToBackendMut<BE> + GLWEInfos,
+        I: GLWEInfos,
+        A: GLWEToBackendRef<BE> + GLWEInfos,
+        B: GLWEToBackendRef<BE> + GLWEInfos,
+        T: GGLWEInfos + GLWETensorKeyPreparedToBackendRef<BE> + GGLWEPreparedToBackendRef<BE>,
+    {
+        let arena = scratch.borrow();
+        let (mut tensor, mut arena) = arena.take_glwe_tensor_scratch(tensor_infos);
+        BE::glwe_tensor_apply(self, cnv_offset, &mut tensor, a, b, &mut arena);
+        BE::glwe_tensor_relinearize(self, res, &tensor, tsk, &mut arena);
+    }
+
+    fn glwe_tensor_square_relinearize_assign_default<R, I, T>(
+        &self,
+        cnv_offset: usize,
+        res: &mut R,
+        tensor_infos: &I,
+        tsk: &T,
+        scratch: &mut ScratchArena<'_, BE>,
+    ) where
+        BE: GLWETensoringImpl<BE>,
+        R: GLWEToBackendMut<BE> + GLWEToBackendRef<BE> + GLWEInfos,
+        I: GLWEInfos,
+        T: GGLWEInfos + GLWETensorKeyPreparedToBackendRef<BE> + GGLWEPreparedToBackendRef<BE>,
+    {
+        let arena = scratch.borrow();
+        let (mut tensor, mut arena) = arena.take_glwe_tensor_scratch(tensor_infos);
+        BE::glwe_tensor_square_apply(self, cnv_offset, &mut tensor, &*res, &mut arena);
+        BE::glwe_tensor_relinearize(self, res, &tensor, tsk, &mut arena);
+    }
+
+    fn glwe_tensor_apply_prepared_right_relinearize_assign_default<R, I, BP, T>(
+        &self,
+        cnv_offset: usize,
+        res: &mut R,
+        tensor_infos: &I,
+        b_prep: &BP,
+        b_size: usize,
+        tsk: &T,
+        scratch: &mut ScratchArena<'_, BE>,
+    ) where
+        BE: GLWETensoringImpl<BE>,
+        R: GLWEToBackendMut<BE> + GLWEToBackendRef<BE> + GLWEInfos,
+        I: GLWEInfos,
+        BP: CnvPVecRToBackendRef<BE>,
+        T: GGLWEInfos + GLWETensorKeyPreparedToBackendRef<BE> + GGLWEPreparedToBackendRef<BE>,
+    {
+        let arena = scratch.borrow();
+        let (mut tensor, mut arena) = arena.take_glwe_tensor_scratch(tensor_infos);
+        BE::glwe_tensor_apply_prepared_right(self, cnv_offset, &mut tensor, &*res, b_prep, b_size, &mut arena);
+        BE::glwe_tensor_relinearize(self, res, &tensor, tsk, &mut arena);
     }
 }