+++ poulpy-core/src/delegates/operations.rs
@@ -266,6 +266,55 @@ impl_operations_delegate!(
         B: GGLWEInfos,
     {
         BE::glwe_tensor_relinearize_tmp_bytes(self, res, a, tsk)
+    },
+    fn glwe_tensor_apply_relinearize<R, I, A, B, T>(
+        &self,
+        cnv_offset: usize,
+        res: &mut R,
+        tensor_infos: &I,
+        a: &A,
+        b: &B,
+        tsk: &T,
+        scratch: &mut ScratchArena<'_, BE>,
+    ) where
+        R: GLWEToBackendMut<BE> + GLWEInfos,
+        I: GLWEInfos,
+        A: GLWEToBackendRef<BE> + GLWEInfos,
+        B: GLWEToBackendRef<BE> + GLWEInfos,
+        T: GGLWEInfos + GLWETensorKeyPreparedToBackendRef<BE> + GGLWEPreparedToBackendRef<BE>,
+    {
+        BE::glwe_tensor_apply_relinearize(self, cnv_offset, res, tensor_infos, a, b, tsk, scratch)
+    },
+    fn glwe_tensor_square_relinearize_assign<R, I, T>(
+        &self,
+        cnv_offset: usize,
+        res: &mut R,
+        tensor_infos: &I,
+        tsk: &T,
+        scratch: &mut ScratchArena<'_, BE>,
+    ) where
+        R: GLWEToBackendMut<BE> + GLWEToBackendRef<BE> + GLWEInfos,
+        I: GLWEInfos,
+        T: GGLWEInfos + GLWETensorKeyPreparedToBackendRef<BE> + GGLWEPreparedToBackendRef<BE>,
+    {
+        BE::glwe_tensor_square_relinearize_assign(self, cnv_offset, res, tensor_infos, tsk, scratch)
+    },
+    fn glwe_tensor_apply_prepared_right_relinearize_assign<R, I, BP, T>(
+        &self,
+        cnv_offset: usize,
+        res: &mut R,
+        tensor_infos: &I,
+        b_prep: &BP,
+        b_size: usize,
+        tsk: &T,
+        scratch: &mut ScratchArena<'_, BE>,
+    ) where
+        R: GLWEToBackendMut<BE> + GLWEToBackendRef<BE> + GLWEInfos,
+        I: GLWEInfos,
+        BP: CnvPVecRToBackendRef<BE>,
+        T: GGLWEInfos + GLWETensorKeyPreparedToBackendRef<BE> + GGLWEPreparedToBackendRef<BE>,
+    {
+        BE::glwe_tensor_apply_prepared_right_relinearize_assign(self, cnv_offset, res, tensor_infos, b_prep, b_size, tsk, scratch)
     }
 );